requirements.txt
flask
python-telegram-bot==20.6
transformers
torch
langdetect